chore: removed SetHost and shifted SetDataResidency to sendgrid.go
