+++ sendgrid.go
@@ -1,7 +1,9 @@
 package sendgrid
 
 import (
+	"errors"
 	"github.com/sendgrid/rest"
+	"net/url"
 )
 
 // sendGridOptions for CreateRequest
@@ -10,6 +12,12 @@ type sendGridOptions struct {
 	Endpoint string
 	Host     string
 	Subuser  string
+}
+
+// sendgrid host map for different regions
+var allowedRegionsHostMap = map[string]string{
+	"eu":     "https://api.eu.sendgrid.com",
+	"global": "https://api.sendgrid.com",
 }
 
 // GetRequest
@@ -46,4 +54,37 @@ func NewSendClient(key string) *Client {
 	request := GetRequest(key, "/v3/mail/send", "")
 	request.Method = "POST"
 	return &Client{request}
+}
+
+// extractEndpoint extracts the endpoint from a baseURL
+func extractEndpoint(link string) (string, error) {
+	parsedURL, err := url.Parse(link)
+	if err != nil {
+		return "", err
+	}
+
+	return parsedURL.Path, nil
+}
+
+// SetDataResidency modifies the host as per the region
+/*
+ * This allows support for global and eu regions only. This set will likely expand in the future.
+ * Global should be the default
+ * Global region means the message should be sent through:
+ * HTTP: api.sendgrid.com
+ * EU region means the message should be sent through:
+ * HTTP: api.eu.sendgrid.com
+ */
+// @return [Request] the modified request object
+func SetDataResidency(request rest.Request, region string) (rest.Request, error) {
+	regionalHost, present := allowedRegionsHostMap[region]
+	if !present {
+		return request, errors.New("error: region can only be \"eu\" or \"global\"")
+	}
+	endpoint, err := extractEndpoint(request.BaseURL)
+	if err != nil {
+		return request, err
+	}
+	request.BaseURL = regionalHost + endpoint
+	return request, nil
 }
